advancedtls: configure tls.GetConfigForClient to update root certs
If tls.RequireAndVerifyClientCert and RootCertificateOptions.RootProvider are
set for NewServerCreds, the verification check built by advancedtls will not
to be executed because it fails the verification check in crypto/tls.

Setting tls.RequireAndVerifyClientCert invokes crypto/tls's verification check.
However, since RootCertificateOptions.RootProvider is set for advancedtls, this
does not set tls.ClientCAs, which is required for crypto/tls's verification
check. So it fails and never invokes advancedtls's verification check.

This fix dynamically reloads the server root certificates tls.ClientCAs using
tls.GetConfigForClient based on the KeyMaterial function provided by
RootProvider so that the verification check for both crypto/tls and advancedtls
is invoked.